requirements.txt
Flask==3.0.0
openai==1.6.1
python-dotenv==1.0.0
python-docx==1.1.0
flask-cors==4.0.0